% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevitour.R
\name{langevitour}
\alias{langevitour}
\title{Langevin Tour}
\usage{
langevitour(X, groups = NULL, width = NULL, height = NULL, elementId = NULL)
}
\description{
Create a Langevin Tour HTML widget. Can be used, for example, from the console or in an RMarkdown document.
}
