% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevitour.R
\name{langevitour}
\alias{langevitour}
\title{Langevin Tour}
\usage{
langevitour(
  X,
  group = NULL,
  center = NULL,
  scale = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{X}{The data to plot. A matrix of numeric data, or something that can be cast to a matrix.}

\item{group}{A group for each row in X, will be used to color points. A factor, or something that can be cast to a factor.}

\item{center}{Center for each variable. If omitted, the column means will be used.}

\item{scale}{Scale for each variable. Scale +/- center will be the range of guaranteed visible data. If omitted, a reasonable default will be chosen.}

\item{width}{Width of widget.}

\item{height}{Height of widget.}

\item{elementId}{An element ID for the widget, see htmlwidgets::createWidget.}
}
\value{
An htmlwidget object.
}
\description{
Create a Langevin Tour HTML widget. Can be used, for example, from the console or in an RMarkdown document.
}
\examples{
langevitour(iris[,1:4], iris$Species)

}
